1. Implement DPOS DAG blocks proposal and verification 2. Refactor DAG block, decoupling DAG block manager from DAG block structure
