+++ src/consensus/block_proposer.hpp
@@ -8,7 +8,7 @@
 
 #include "boost/thread.hpp"
 #include "config/config.hpp"
-#include "dag/dag_block.hpp"
+#include "dag/dag_block_manager.hpp"
 #include "dag/vdf_sortition.hpp"
 #include "logger/log.hpp"
 #include "network/network.hpp"
@@ -72,11 +72,11 @@ class BlockProposer : public std::enable_shared_from_this<BlockProposer> {
  public:
   BlockProposer(BlockProposerConfig const& bp_config, vdf_sortition::VdfConfig const& vdf_config,
                 std::shared_ptr<DagManager> dag_mgr, std::shared_ptr<TransactionManager> trx_mgr,
-                std::shared_ptr<BlockManager> blk_mgr, std::shared_ptr<FinalChain> final_chain, addr_t node_addr,
+                std::shared_ptr<DagBlockManager> dag_blk_mgr, std::shared_ptr<FinalChain> final_chain, addr_t node_addr,
                 secret_t node_sk, vrf_sk_t vrf_sk, logger::Logger log_time)
       : dag_mgr_(dag_mgr),
         trx_mgr_(trx_mgr),
-        blk_mgr_(blk_mgr),
+        dag_blk_mgr_(dag_blk_mgr),
         final_chain_(final_chain),
         bp_config_(bp_config),
         log_time_(log_time),
@@ -119,7 +119,7 @@ class BlockProposer : public std::enable_shared_from_this<BlockProposer> {
   uint16_t my_trx_shard_;
   std::shared_ptr<DagManager> dag_mgr_;
   std::shared_ptr<TransactionManager> trx_mgr_;
-  std::shared_ptr<BlockManager> blk_mgr_;
+  std::shared_ptr<DagBlockManager> dag_blk_mgr_;
   std::shared_ptr<FinalChain> final_chain_;
   std::shared_ptr<std::thread> proposer_worker_;
   std::unique_ptr<ProposeModelFace> propose_model_;
