+++ src/dag/vdf_sortition.cpp
@@ -1,7 +1,9 @@
 #include "vdf_sortition.hpp"
 
-#include "config/config.hpp"  // The order is important
-#include "libdevcore/CommonData.h"
+#include <libdevcore/CommonData.h>
+#include <libdevcore/CommonJS.h>
+
+#include "config/config.hpp"
 
 namespace taraxa::vdf_sortition {
