+++ src/node/full_node.cpp
@@ -97,18 +97,19 @@ void FullNode::init() {
   auto genesis_hash = conf_.chain.dag_genesis_block.getHash().toString();
   emplace(pbft_chain_, genesis_hash, node_addr, db_);
   emplace(dag_mgr_, genesis_hash, node_addr, trx_mgr_, pbft_chain_, db_);
-  emplace(blk_mgr_, conf_.chain.vdf, 1024 /*capacity*/, 4 /* verifer thread*/, node_addr, db_, trx_mgr_, log_time_,
-          conf_.test_params.max_block_queue_warn);
+  emplace(dag_blk_mgr_, conf_.chain.vdf, 1024 /*capacity*/, 4 /* verifer thread*/, node_addr, db_, trx_mgr_,
+          final_chain_, pbft_chain_, log_time_, conf_.test_params.max_block_queue_warn);
   emplace(vote_mgr_, node_addr, final_chain_, pbft_chain_);
   emplace(trx_order_mgr_, node_addr, db_);
   emplace(executor_, node_addr, db_, dag_mgr_, trx_mgr_, final_chain_, pbft_chain_,
           conf_.test_params.block_proposer.transaction_limit);
-  emplace(pbft_mgr_, conf_.chain.pbft, genesis_hash, node_addr, db_, pbft_chain_, vote_mgr_, dag_mgr_, blk_mgr_,
+  emplace(pbft_mgr_, conf_.chain.pbft, genesis_hash, node_addr, db_, pbft_chain_, vote_mgr_, dag_mgr_, dag_blk_mgr_,
           final_chain_, executor_, kp_.secret(), conf_.vrf_secret);
-  emplace(blk_proposer_, conf_.test_params.block_proposer, conf_.chain.vdf, dag_mgr_, trx_mgr_, blk_mgr_, final_chain_,
-          node_addr, getSecretKey(), getVrfSecretKey(), log_time_);
+  emplace(blk_proposer_, conf_.test_params.block_proposer, conf_.chain.vdf, dag_mgr_, trx_mgr_, dag_blk_mgr_,
+          final_chain_, node_addr, getSecretKey(), getVrfSecretKey(), log_time_);
   emplace(network_, conf_.network, conf_.net_file_path().string(), kp_.secret(), genesis_hash, node_addr, db_,
-          pbft_mgr_, pbft_chain_, vote_mgr_, dag_mgr_, blk_mgr_, trx_mgr_, kp_.pub(), conf_.chain.pbft.lambda_ms_min);
+          pbft_mgr_, pbft_chain_, vote_mgr_, dag_mgr_, dag_blk_mgr_, trx_mgr_, kp_.pub(),
+          conf_.chain.pbft.lambda_ms_min);
 
   // Inits rpc related members
   if (conf_.rpc) {
@@ -164,11 +165,11 @@ void FullNode::start() {
   executor_->start();
   pbft_mgr_->setNetwork(network_);
   pbft_mgr_->start();
-  blk_mgr_->start();
+  dag_blk_mgr_->start();
   block_workers_.emplace_back([this]() {
     while (!stopped_) {
       // will block if no verified block available
-      auto blk = blk_mgr_->popVerifiedBlock();
+      auto blk = dag_blk_mgr_->popVerifiedBlock();
 
       if (!stopped_) {
         received_blocks_++;
@@ -187,10 +188,10 @@ void FullNode::start() {
         // where in some race condition older block is verfified faster then
         // new block but should resolve quickly, return block to queue
         if (!stopped_) {
-          if (blk_mgr_->pivotAndTipsValid(blk)) {
+          if (dag_blk_mgr_->pivotAndTipsValid(blk)) {
             LOG(log_dg_) << "Block could not be added to DAG " << blk.getHash().toString();
             received_blocks_--;
-            blk_mgr_->pushVerifiedBlock(blk);
+            dag_blk_mgr_->pushVerifiedBlock(blk);
           }
         }
       }
@@ -236,7 +237,7 @@ void FullNode::close() {
   trx_mgr_->stop();
   trx_mgr_->setNetwork(nullptr);
   network_->stop();
-  blk_mgr_->stop();
+  dag_blk_mgr_->stop();
   for (auto &t : block_workers_) {
     t.join();
   }
@@ -281,12 +282,12 @@ void FullNode::rebuildDb() {
         auto dag_block = old_db_->getDagBlock(dag_block_hash);
         auto pivot_hash = dag_block->getPivot();
         auto tips = dag_block->getTips();
-        if (pbft_dag_blocks.count(pivot_hash) == 0 && !(blk_mgr_->isBlockKnown(pivot_hash))) {
+        if (pbft_dag_blocks.count(pivot_hash) == 0 && !(dag_blk_mgr_->isBlockKnown(pivot_hash))) {
           pbft_dag_blocks.emplace(pivot_hash);
           new_dag_blocks.push_back(pivot_hash);
         }
         for (auto &tip : tips) {
-          if (pbft_dag_blocks.count(tip) == 0 && !(blk_mgr_->isBlockKnown(tip))) {
+          if (pbft_dag_blocks.count(tip) == 0 && !(dag_blk_mgr_->isBlockKnown(tip))) {
             pbft_dag_blocks.emplace(tip);
             new_dag_blocks.push_back(tip);
           }
@@ -324,7 +325,7 @@ void FullNode::rebuildDb() {
       for (auto const &block : block_level.second) {
         LOG(log_nf_) << "Storing block " << block.second.first.getHash().toString() << " with "
                      << block.second.second.size() << " transactions";
-        blk_mgr_->insertBroadcastedBlockWithTransactions(block.second.first, block.second.second);
+        dag_blk_mgr_->insertBroadcastedBlockWithTransactions(block.second.first, block.second.second);
       }
     }
