+++ src/transaction_manager/transaction_manager.hpp
@@ -14,6 +14,7 @@ namespace taraxa {
 
 using std::string;
 class DagBlock;
+class DagManager;
 class FullNode;
 class Network;
 namespace net {
