+++ tests/transaction_test.cpp
@@ -1,6 +1,7 @@
 #include "transaction_manager/transaction.hpp"
 
 #include <gtest/gtest.h>
+#include <libdevcore/CommonJS.h>
 
 #include <thread>
 #include <vector>
